# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MLBB Collector
  version: 1.0.0
  description: >
    Mobile Legends: Bang Bang heroes, counters, and synergies, with AI scoring.
    Fan-made; not affiliated with Moonton.
tags:
  - name: health
  - name: heroes
    description: Heroes, their counters, and their synergies.
  - name: counters
    description: AI counter scoring.
  - name: synergies
    description: AI synergy scoring.

paths:
  /health:
    get:
      tags: [health]
      summary: Liveness
      operationId: health
      responses:
        "200":
          description: The service is up.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: { type: string, const: ok }

  /api/heroes:
    get:
      tags: [heroes]
      summary: List heroes
      description: >
        Ordered by Moonton ID. An invalid `page` or `size` uses the default;
        `size` is capped at 100.
      operationId: listHeroes
      parameters:
        - name: search
          in: query
          description: Name contains this text, case-insensitive.
          schema: { type: string }
        - name: role
          in: query
          description: Has this role, case-insensitive.
          schema: { $ref: "#/components/schemas/Role" }
        - name: lane
          in: query
          description: Plays this lane, case-insensitive.
          schema: { $ref: "#/components/schemas/Lane" }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: size
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
      responses:
        "200":
          description: One page of heroes.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HeroPage" }

  /api/heroes/{heroId}:
    get:
      tags: [heroes]
      summary: Get a hero
      operationId: getHero
      parameters:
        - $ref: "#/components/parameters/HeroId"
      responses:
        "200":
          description: The hero.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Hero" }
        "404":
          $ref: "#/components/responses/HeroNotFound"

  /api/heroes/{heroId}/counters:
    get:
      tags: [heroes]
      summary: List a hero's counters
      description: Ordered by counter hero ID.
      operationId: listCounters
      parameters:
        - $ref: "#/components/parameters/HeroId"
      responses:
        "200":
          description: The hero's counters.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Counter" }
        "404":
          description: "`hero_not_found` or `counter_data_not_found`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /api/heroes/{heroId}/synergies:
    get:
      tags: [heroes]
      summary: List a hero's synergies
      description: Ordered by synergy hero ID.
      operationId: listSynergies
      parameters:
        - $ref: "#/components/parameters/HeroId"
      responses:
        "200":
          description: The hero's synergies.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Synergy" }
        "404":
          description: "`hero_not_found` or `synergy_data_not_found`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /api/counters/analyze-score:
    post:
      tags: [counters]
      summary: Score every counter of a target hero
      operationId: analyzeCounterScore
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                targetHeroId: { type: string, examples: [tigreal] }
                language: { $ref: "#/components/schemas/Language" }
      responses:
        "200":
          description: Ranked by score, then confidence, then hero ID.
          headers:
            Set-Cookie: { $ref: "#/components/headers/ClientCookie" }
          content:
            application/json:
              schema:
                type: object
                required: [targetHeroId, source, recommendations]
                properties:
                  targetHeroId: { type: string }
                  source: { type: string, const: ai }
                  recommendations:
                    type: array
                    items:
                      type: object
                      required: [rank, counterHeroId, score, confidence]
                      properties:
                        rank: { type: integer, minimum: 1 }
                        counterHeroId: { type: string }
                        score: { $ref: "#/components/schemas/Rating" }
                        confidence: { $ref: "#/components/schemas/Rating" }
        "404":
          description: "`target_hero_not_found` or `counter_data_not_found`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422": { $ref: "#/components/responses/InvalidRequest" }
        "429": { $ref: "#/components/responses/RateLimitExceeded" }
        "502": { $ref: "#/components/responses/ProviderError" }
        "503": { $ref: "#/components/responses/RateLimitUnavailable" }
        "504": { $ref: "#/components/responses/ProviderUnavailable" }

  /api/counters/analyze-detail:
    post:
      tags: [counters]
      summary: Explain one counter
      operationId: analyzeCounterDetail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                targetHeroId: { type: string, examples: [tigreal] }
                counterHeroId: { type: string, examples: [diggie] }
                language: { $ref: "#/components/schemas/Language" }
      responses:
        "200":
          description: The counter's score and explanation.
          headers:
            Set-Cookie: { $ref: "#/components/headers/ClientCookie" }
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    required: [targetHeroId, counterHeroId]
                    properties:
                      targetHeroId: { type: string }
                      counterHeroId: { type: string }
                  - $ref: "#/components/schemas/Detail"
        "404":
          description: "`target_hero_not_found`, `counter_data_not_found`, `counter_hero_not_found`, or `counter_matchup_not_found`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422": { $ref: "#/components/responses/InvalidRequest" }
        "429": { $ref: "#/components/responses/RateLimitExceeded" }
        "502": { $ref: "#/components/responses/ProviderError" }
        "503": { $ref: "#/components/responses/RateLimitUnavailable" }
        "504": { $ref: "#/components/responses/ProviderUnavailable" }

  /api/synergies/analyze-score:
    post:
      tags: [synergies]
      summary: Score every synergy of an anchor hero
      operationId: analyzeSynergyScore
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                anchorHeroId: { type: string, examples: [tigreal] }
                language: { $ref: "#/components/schemas/Language" }
      responses:
        "200":
          description: Ranked by score, then confidence, then hero ID.
          headers:
            Set-Cookie: { $ref: "#/components/headers/ClientCookie" }
          content:
            application/json:
              schema:
                type: object
                required: [anchorHeroId, source, recommendations]
                properties:
                  anchorHeroId: { type: string }
                  source: { type: string, const: ai }
                  recommendations:
                    type: array
                    items:
                      type: object
                      required: [rank, synergyHeroId, score, confidence]
                      properties:
                        rank: { type: integer, minimum: 1 }
                        synergyHeroId: { type: string }
                        score: { $ref: "#/components/schemas/Rating" }
                        confidence: { $ref: "#/components/schemas/Rating" }
        "404":
          description: "`anchor_hero_not_found` or `synergy_data_not_found`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422": { $ref: "#/components/responses/InvalidRequest" }
        "429": { $ref: "#/components/responses/RateLimitExceeded" }
        "502": { $ref: "#/components/responses/ProviderError" }
        "503": { $ref: "#/components/responses/RateLimitUnavailable" }
        "504": { $ref: "#/components/responses/ProviderUnavailable" }

  /api/synergies/analyze-detail:
    post:
      tags: [synergies]
      summary: Explain one synergy
      operationId: analyzeSynergyDetail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                anchorHeroId: { type: string, examples: [tigreal] }
                synergyHeroId: { type: string, examples: [pharsa] }
                language: { $ref: "#/components/schemas/Language" }
      responses:
        "200":
          description: The synergy's score and explanation.
          headers:
            Set-Cookie: { $ref: "#/components/headers/ClientCookie" }
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    required: [anchorHeroId, synergyHeroId]
                    properties:
                      anchorHeroId: { type: string }
                      synergyHeroId: { type: string }
                  - $ref: "#/components/schemas/Detail"
        "404":
          description: "`anchor_hero_not_found`, `synergy_data_not_found`, `synergy_hero_not_found`, or `synergy_matchup_not_found`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422": { $ref: "#/components/responses/InvalidRequest" }
        "429": { $ref: "#/components/responses/RateLimitExceeded" }
        "502": { $ref: "#/components/responses/ProviderError" }
        "503": { $ref: "#/components/responses/RateLimitUnavailable" }
        "504": { $ref: "#/components/responses/ProviderUnavailable" }

components:
  parameters:
    HeroId:
      name: heroId
      in: path
      required: true
      description: Hero slug.
      schema: { type: string, examples: [tigreal] }

  schemas:
    Role:
      type: string
      enum: [tank, fighter, assassin, mage, marksman, support]
    Lane:
      type: string
      enum: [gold, exp, mid, roam, jungle]
    Language:
      type: string
      enum: [en, id]
      default: en
    Rating:
      type: integer
      minimum: 0
      maximum: 100

    Hero:
      type: object
      required: [uid, mlid, name, roles, lanes, images]
      properties:
        uid: { type: string, description: Hero slug., examples: [tigreal] }
        mlid: { type: string, pattern: "^[1-9][0-9]*$", description: The number Moonton assigns the hero, as a string., examples: ["6"] }
        name: { type: string, examples: [Tigreal] }
        roles:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/Role" }
        lanes:
          type: array
          items: { $ref: "#/components/schemas/Lane" }
        images:
          type: object
          description: Image URLs by variant (`head`).
          additionalProperties: { type: string }

    HeroPage:
      type: object
      required: [items, page, size, total, pages]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Hero" }
        page: { type: integer, minimum: 1 }
        size: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer, minimum: 0 }
        pages: { type: integer, minimum: 0 }

    ProofBase:
      type: object
      required: [id, category, priority, impact, summary, worksBestWhen, failureCases]
      properties:
        id: { type: string }
        priority: { type: string, enum: [primary, secondary, condition] }
        impact: { type: string, enum: [high, medium, low] }
        summary: { type: string }
        worksBestWhen:
          type: array
          minItems: 1
          items: { type: string }
        failureCases:
          type: array
          minItems: 1
          items: { type: string }

    CounterProof:
      allOf:
        - $ref: "#/components/schemas/ProofBase"
        - type: object
          properties:
            category:
              type: string
              enum: [skill-interaction, crowd-control-counter, damage-type-advantage, item-power-spike, mobility-advantage, range-advantage, kiting, sustain-anti-sustain, positioning-requirement, cooldown-window, vision-awareness, teamfight-role-counter, game-phase, execution-difficulty]

    SynergyProof:
      allOf:
        - $ref: "#/components/schemas/ProofBase"
        - type: object
          properties:
            category:
              type: string
              enum: [skill-interaction, crowd-control-chain, engage-follow-up, setup-combo, damage-amplification, protection, peel, frontline-enabler, mobility-enabler, vision-setup, healing-sustain, shielding, poke-siege, pickoff-combo, teamfight-combo, objective-control, laning-synergy, game-phase, positioning-requirement, cooldown-window, execution-difficulty]

    Counter:
      type: object
      required: [targetHeroId, counterHero, reasons, counterTypes, proof]
      properties:
        targetHeroId: { type: string, examples: [tigreal] }
        counterHero: { $ref: "#/components/schemas/Hero" }
        reasons:
          type: array
          minItems: 1
          items: { type: string }
        counterTypes:
          type: array
          minItems: 1
          items: { type: string }
        proof:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/CounterProof" }

    Synergy:
      type: object
      required: [anchorHeroId, synergyHero, reasons, synergyTypes, proof]
      properties:
        anchorHeroId: { type: string, examples: [tigreal] }
        synergyHero: { $ref: "#/components/schemas/Hero" }
        reasons:
          type: array
          minItems: 1
          items: { type: string }
        synergyTypes:
          type: array
          minItems: 1
          items: { type: string }
        proof:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/SynergyProof" }

    Detail:
      type: object
      required: [source, score, confidence, summary, strengths, conditions, failureCases, evidenceIds]
      properties:
        source: { type: string, const: ai }
        score: { $ref: "#/components/schemas/Rating" }
        confidence: { $ref: "#/components/schemas/Rating" }
        summary: { type: string }
        strengths:
          type: array
          minItems: 1
          items: { type: string }
        conditions:
          type: array
          items: { type: string }
        failureCases:
          type: array
          items: { type: string }
        evidenceIds:
          type: array
          description: Proof IDs of this counter or synergy that the explanation draws on.
          items: { type: string }

    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, examples: [hero_not_found] }
            message: { type: string, examples: [Hero was not found in the dataset.] }

  responses:
    HeroNotFound:
      description: "`hero_not_found`."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    InvalidRequest:
      description: "`invalid_request`: the body is not valid JSON, or `language` is not `en` or `id`."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimitExceeded:
      description: "`rate_limit_exceeded`: the address or browser has used its quota for this route."
      headers:
        Retry-After:
          description: Seconds until the quota resets.
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimitUnavailable:
      description: "`rate_limit_unavailable`: rate limiting is enabled and Redis cannot answer."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ProviderError:
      description: "`ai_provider_error`: every provider failed, or the model's answer was invalid."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ProviderUnavailable:
      description: "`ai_provider_not_configured` or `ai_provider_timeout`."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }

  headers:
    ClientCookie:
      description: The browser's HttpOnly rate-limit ID, set when rate limiting is enabled and the request carried no valid one.
      schema: { type: string, examples: ["mlbb_collector_client_id=6f1c2a54-0d6e-4b8a-9a7e-3c5d1f2e8b90; Path=/; Max-Age=2592000; HttpOnly; SameSite=Lax"] }